+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - [builder] Removed useless download.kiwix.org special behavior for reader checksum. Checksum now on Reader
+- [builder] Using captive-portal 1.4.1
 - [inputs.base] Version-only base image def now targets uncompressed URL
 - [inputs.base] BaseConfig gets an optional `checksum` and populates base_file accordingly
 - [inputs.base] Version-only base image def now also retrieves Checksum via expected .md5 endpoint